fix: only delete files if not on CI

+++ node/logger.js
@@ -28,8 +28,10 @@ function ico(type) {
 }
 
 exports.init = () => {
-  storage.delete(GB_DECOR)
-  storage.delete(LOG_PATH)
+  if (!system.isCI()) {
+    storage.delete(GB_DECOR)
+    storage.delete(LOG_PATH)
+  }
   storage.makeDir(LOG_PATH)
 }
